requirements.txt
bokeh==3.0.3
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
imageio==2.25.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.6.3
networkx==3.0
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
scikit-image==0.19.3
scipy==1.10.0
six==1.16.0
tifffile==2023.1.23.1
tornado==6.2
xyzservices==2022.9.0
